3-ordersystem-microservices-distributed-system/address-api/requirements.txt
pylint==2.8.3 
flask==2.0.1
Flask-Script==2.0.6
coverage==5.5
cassandra-driver==3.25.0
uwsgi==2.0.19.1
Werkzeug==2.0.1